# services/control-plane/internal/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velane Control Plane API",
    "version": "1.0.0",
    "description": "HTTP API for Velane authentication, tenant administration, workflows, integrations, invocation, observability, and embeds. Unless stated otherwise, authenticated operations accept either a Velane API key (`vl_...`), a session JWT, or the session cookie. The `x-required-scope` extension records the minimum API-key scope or session role."
  },
  "servers": [
    { "url": "/", "description": "Current Velane deployment" }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "sessionCookie": [] }
  ],
  "paths": {
...
    "/v1/tenants": {
      "post": { "operationId": "createTenant", "summary": "Bootstrap a tenant", "description": "Intentionally unauthenticated for self-hosted bootstrap. Production deployments should restrict this endpoint at the network edge.", "tags": ["Tenants"], "security": [], "requestBody": { "$ref": "#/components/requestBodies/CreateTenant" }, "responses": { "201": { "description": "Tenant created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tenant" } } } }, "400": { "$ref": "#/components/responses/BadRequest" } } }
    },
...
    "/v1/tenant/api-keys": {
      "get": { "operationId": "listAPIKeys", "summary": "List API keys", "tags": ["Tenants"], "x-required-scope": "admin", "responses": { "200": { "description": "API keys without secret values", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/APIKey" } } } } }, "403": { "$ref": "#/components/responses/Forbidden" } } },
      "post": { "operationId": "createAPIKey", "summary": "Create an API key", "tags": ["Tenants"], "x-required-scope": "admin", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAPIKeyRequest" } } } }, "responses": { "201": { "description": "Created API key; the secret is returned only once", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/APIKey" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "403": { "$ref": "#/components/responses/Forbidden" } } }
    },
...
    "/v1/tenant/egress": {
      "get": { "operationId": "getEgressPolicy", "summary": "Get the tenant egress policy", "tags": ["Tenants"], "responses": { "200": { "description": "Egress policy", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EgressPolicy" } } } } } },
      "put": { "operationId": "updateEgressPolicy", "summary": "Update the tenant egress policy", "tags": ["Tenants"], "x-required-scope": "admin", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EgressPolicy" } } } }, "responses": { "200": { "description": "Updated egress policy", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EgressPolicy" } } } }, "400": { "$ref": "#/components/responses/BadRequest" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Velane API key (`vl_...`) or RS256 session JWT." },
      "sessionBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "RS256 admin session JWT. API keys are not accepted by session-only endpoints." },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "velane_session", "description": "HttpOnly admin session cookie." },
      "embedToken": { "type": "http", "scheme": "bearer", "description": "Opaque embed token (`et_...`). Embed tokens never receive admin scope." },
      "gitWebhookSignature": { "type": "apiKey", "in": "header", "name": "X-Hub-Signature-256", "description": "HMAC-SHA256 signature in `sha256=<hex>` format." },
      "nangoWebhookSignature": { "type": "apiKey", "in": "header", "name": "X-Nango-Hmac-Sha256", "description": "Nango webhook HMAC signature." }
    },
...
    "requestBodies": {
      "CreateTenant": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTenantRequest" } } } },
...
    },
    "responses": {
...
      "BadRequest": { "description": "Invalid request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Forbidden": { "description": "The credential lacks access or the required scope.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } } },
...
      "CreateTenantRequest": { "type": "object", "required": ["name", "slug"], "properties": { "name": { "type": "string", "minLength": 1 }, "slug": { "type": "string", "pattern": "^[a-z0-9][a-z0-9-]{1,61}[a-z0-9]$" } } },
      "Tenant": { "type": "object", "required": ["id", "name", "slug", "created_at", "egress_policy", "replay_enabled", "branding", "runtime_limits"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "slug": { "type": "string" }, "created_at": { "type": "string", "format": "date-time" }, "egress_policy": { "$ref": "#/components/schemas/EgressPolicy" }, "replay_enabled": { "type": "boolean" }, "branding": { "$ref": "#/components/schemas/Branding" }, "runtime_limits": { "$ref": "#/components/schemas/RuntimeLimits" }, "license_key": { "type": ["string", "null"] } } },
...
      "Branding": { "type": "object", "properties": { "logo_url": { "type": "string", "format": "uri" }, "accent_color": { "type": "string" }, "font_family": { "type": "string" }, "custom_domain": { "type": "string" }, "hide_branding": { "type": "boolean" } } },
      "EgressPolicy": { "type": "object", "required": ["blocked_cidrs", "blocked_domains"], "properties": { "blocked_cidrs": { "type": "array", "items": { "type": "string" } }, "blocked_domains": { "type": "array", "items": { "type": "string" } } } },
      "RuntimeLimits": { "type": "object", "required": ["max_timeout_ms", "max_memory_mb", "max_cpu_percent"], "properties": { "max_timeout_ms": { "type": "integer" }, "max_memory_mb": { "type": "integer" }, "max_cpu_percent": { "type": "integer" } } },
      "CreateAPIKeyRequest": { "type": "object", "required": ["name", "scopes"], "properties": { "name": { "type": "string" }, "scopes": { "type": "array", "items": { "type": "string", "enum": ["invoke", "manage", "admin"] }, "uniqueItems": true } } },
      "APIKey": { "type": "object", "required": ["id", "name", "scopes", "key_prefix", "created_at"], "properties": { "id": { "type": "string" }, "tenant_id": { "type": "string", "description": "Present in the create response." }, "name": { "type": "string" }, "scopes": { "type": "array", "items": { "type": "string" } }, "key_prefix": { "type": "string" }, "key": { "type": "string", "pattern": "^vl_", "description": "Present only in the create response." }, "expires_at": { "type": ["string", "null"], "format": "date-time" }, "last_used_at": { "type": ["string", "null"], "format": "date-time" }, "created_at": { "type": "string", "format": "date-time" } } },
...
    }
  }
}